Remove dashboard info from single entry view
A GravityView and all of its configurations are displayed in whatever page template the shortcode is added to — in this case, the dashboard template. However, a single entry has its own layout and should not be displayed with the main dashboard information.

This commit adjusts the dashboard template to only show the the dashboard information if a single entry is not being displayed, using the GravityView query variable.